// frontend/src/apps/transactions/models/base/utils/baseSchema.ts
/* LastChecked: 2026-03-14 | WhereUsed: TODO(wc3-schema-audit) | WhoCreated: Unknown */
import * as z from "zod";

export const baseTransactionSchema = z.object({
  id_transaction: z.string().optional(),
  dt_created: z.number().optional(),
  dt_updated: z.number().optional(),
  customer_id: z.number().min(1, "Customer ID is required"),
  total: z.number(),
  tax: z.number(),
  discount: z.number(),
  metadata: z.any().optional(),
  prefs: z.any().optional(),
  refs: z.any().optional(),
});